backend-fastapi/worker/requirements.txt
# Worker service dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
PyPDF2==3.0.1
firebase-admin==6.2.0
google-cloud-storage==2.10.0
google-cloud-aiplatform==1.38.1
